Refine public api output
Use cherry-picked parts of the mypy serialization of nodes to generate
more specific public api output.

Namely the variable names of functions, destructured type information and
mro are now included in the output.

There may be other essential fields that haven't yet been thought of
that should be included as well.